requirements.txt
pandas==1.2.0
sklearn==0
seaborn==0.11.1
click==8.1.3
pandas-profiling==3.4.0
marshmallow==3.18.0
marshmallow-dataclass==8.5.9
PyYAML==6.0
numpy==1.21.6